+++ config/coverage.js
@@ -1,0 +1,5 @@
+module.exports = {
+  excludes: [
+    'tests/dummy/**/*'
+  ]
+}