+++ tests/dummy/app/app.js
@@ -1,8 +1,8 @@
 import Ember from 'ember'
 const {Application} = Ember
+import loadInitializers from 'ember-load-initializers'
 
 import config from './config/environment'
-import loadInitializers from 'ember-load-initializers'
 import Resolver from './resolver'
 
 let App