+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# 1.1.2
+
+* **Added** pull request template for Github.
+* **Added** better reporter for test runs.
+* **Added** Firefox to CI build to help ensure we run into no Firefox specific issues.
+* **Added** additional `ember-prop-types` config to make sure no errors slip through the cracks.
+
+# 1.1.1
+
+* **Upgraded** to test against Ember 2.11.
+
+
+# 1.1.0
+
+* **Added** additional builds to CI to make sure addon works with latest versions of Ember.
+* **Removed** files from npm package that aren't necessary (all of the various config files).
+* **Updated** dependencies to latest versions.
+
+
+# 1.0.0
+
+* Dynamic spread is now driven via `spreadOptions=(hash source=(hash object=... property=...))`
+* Added tests
+
+# 0.0.8
+
+* Demo advanced options (action binding, computed property chains)
+
 # 0.0.7
 
 * `till` -> `to`